Change ur_bool_t to uint32_t
This to match the boolean type used in SYCL, where pi_bool is
also uint32_t.

Resolves: #555

+++ include/ur.py
@@ -50,7 +50,7 @@ def UR_MINOR_VERSION( _ver ):
 
 ###############################################################################
 ## @brief compiler-independent type
-class ur_bool_t(c_ubyte):
+class ur_bool_t(c_ulong):
     pass
 
 ###############################################################################
